Rename runtime/utils to runtime/paths
