+++ exonum-java-binding/core/rust/src/runtime/mod.rs
@@ -18,9 +18,9 @@ mod cmd;
 mod config;
 mod java_service_runtime;
 mod service_factory_adapter;
-mod utils;
+mod paths;
 
 pub use self::config::*;
 pub use self::java_service_runtime::JavaServiceRuntime;
 pub use self::service_factory_adapter::JavaServiceFactoryAdapter;
-pub use self::utils::panic_if_java_options;
+pub use self::paths::panic_if_java_options;
